Rename auxilary a python modules for behave
There are two modules with fixtures and Allure formatter. The directory
name was renamed to ci with usages.

(cherry picked from commit de20d86ea2c93f345e34d959e455e08983d1ec19)
