+++ arenadata/scripts/run_behave_tests.bash
@@ -65,7 +65,7 @@ run_feature() {
 
   docker compose -p $project -f arenadata/docker-compose.yaml exec -T \
     -e FEATURE="$feature" -e BEHAVE_FLAGS="--tags $feature --tags=$cluster \
-      -f behave_utils.arenadata.formatter:CustomFormatter \
+      -f behave_utils.ci.formatter:CustomFormatter \
       -o non-existed-output \
       -f allure_behave.formatter:AllureFormatter \
       -o /tmp/allure-results"  \
